requirements.txt
matplotlib==3.4.3
nni==2.9
numpy==1.22.1
PyYAML==6.0
torch==1.9.0
tqdm==4.62.3
